+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [5403](https://github.com/vegaprotocol/vega/issues/5403) - Allow spam free / proof of work free running of null blockchain
 - [5175](https://github.com/vegaprotocol/vega/issues/5175) - Validation free transactions (including signature verification) for null blockchain
 - [5371](https://github.com/vegaprotocol/vega/issues/5371) - Ensure threshold is not breached in ERC20 withdrawal
+- [5358](https://github.com/vegaprotocol/vega/issues/5358) - Update equity shares following updated spec.
 
 ### 🐛 Fixes
 - [5362](https://github.com/vegaprotocol/vega/issues/5362) - Liquidity and order book point to same underlying order after restore